Added command line arguments + updated code
I've added some command line arguments to your program using the click library so the user can specify the number of times the program runs and how long the interval is between each program execution via the command line.

So now for example they could run:
TWD5.py --count 3 --delay 23
Which would check the cookbook page every twenty seconds for three iterations.

I also fixed a weird error which was probably caused by the webpage updating its format.  The previous program kept giving me an error whenever I ran it because it couldn't find the span element with the class "in_stock" because the element didn't exist due to the cookbook not being in stock.  I've changed that so now it checks the parent element (a paragraph element with the class "availability") and returns the text value of the following element which will either be a span with the class "in_stock" or "out_of_stock".

Also I added a snippet of code at the top of the program which tries to install the click library if the user hasn't installed it already.